Create additive.py
Additive Ciphering with encryption and decryption
Here is the implementation of Additive Ciphering. It includes encryption and decryption. Here, decryption is only allowed when the user intends to decrypt it. Here, key should be monoalphabetic (if not, the code will take care of it). Key can be either a number or an alphabet. This code ignores all spaces in the plain text. This code works fine for lower and uppercase alphabets. This code is only for plain text containing 26 alphabets.