Update caesar.py
Caesar Ciphering with encryption and decryption
Here is the implementation of Caesar Ciphering. It includes encryption and decryption. Here, decryption is only allowed when the user intends to decrypt it. Here, key is 3, therefore it's hard-coded within the code itself. This code ignores all spaces in the plain text. This code works fine for lower and uppercase alphabets. This code is only for plain text containing 26 alphabets.